[CDAP-12274] Shows all import errors on Add Entity modal instead of in Studio - Adds error handling if json string from local storage is null or malformed - Moves NonStorePipelineErrorFactory and GLOBALS to react side, then modifies angular service to get from window object - Does validation and show errors on Add Entity modal - Changes styling of Add Entity modal so that modal footer error is still visible on smaller screens
