+++ cdap-ui/app/cdap/components/ResourceCenter/ResourceCenter.scss
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 Cask Data, Inc.
+ * Copyright © 2016-2017 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -18,5 +18,4 @@
   height: 100%;
   padding: 10px;
   overflow-y: auto;
-  max-height: 75vh;
 }
