+++ cdap-ui/app/login/components/Card/index.js
@@ -33,7 +33,8 @@
     </Card>
 */
 
-import React, {Component, PropTypes} from 'react';
+import PropTypes from 'prop-types';
+import React, {Component} from 'react';
 require('./Card.scss');
 
 var classNames = require('classnames');
