+++ cdap-ui/app/login/components/CardActionFeedback/index.js
@@ -22,7 +22,8 @@
       extendedMessage='stack trace message that will get rendered in <pre>'
     />
 */
-import React, {Component, PropTypes} from 'react';
+import PropTypes from 'prop-types';
+import React, {Component} from 'react';
 require('./CardActionFeedback.scss');
 
 var classNames = require('classnames');
